+++ Particles/BlasterImpactParticle.cpp
@@ -7,8 +7,8 @@
 #include "Particles/BlasterImpactParticle.h"
 #include "math.h"
 
-const float startingSize = 0.3;
-const float minLife = 0.5; // Seconds
+const float startingSize = 0.3f;
+const float minLife = 0.5f; // Seconds
 
 BlasterImpactParticle::BlasterImpactParticle(Point3D* _position,
       Vector3D* _velocity, float _life, float _r, float _g, float _b) :
@@ -27,8 +27,8 @@ void BlasterImpactParticle::update(double timeDifference) {
 
 void BlasterImpactParticle::Add(Point3D* pos, Vector3D* vec) {
 
-   float _fade = (0.5 * randdouble()) + minLife;
-   float _r = 0.9 + (randdouble() * 0.1) ;
+   float _fade = (float) ((0.5 * randdouble()) + minLife);
+   float _r = (float) (0.9 + (randdouble() * 0.1));
    float _g = 0;
    float _b = 0;
