+++ Particles/TractorAttractionParticle.cpp
@@ -7,8 +7,8 @@
 #include "Particles/TractorAttractionParticle.h"
 #include "math.h"
 
-const float startingSize = 0.3;
-const float minLife = 0.5; // Seconds
+const float startingSize = 0.3f;
+const float minLife = 0.5f; // Seconds
 TractorAttractionParticle::TractorAttractionParticle(Point3D* _position,
       Vector3D* _velocity, float _life, float _r, float _g, float _b, Point3D* _target) :
    Particle(_position, _velocity, _life, _r, _g, _b) {
@@ -30,9 +30,9 @@ void TractorAttractionParticle::update(double timeDifference) {
 }
 
 void TractorAttractionParticle::Add(Point3D* pos, Vector3D* vec, Point3D* _target) {
-   float _fade = (0.5 * randdouble()) + minLife;
-   float _r = 0.9 + (randdouble() * 0.1);
-   float _g = 0.9 + (randdouble() * 0.1);
+   float _fade = (float) ((0.5 * randdouble()) + minLife);
+   float _r = (float) (0.9 + (randdouble() * 0.1));
+   float _g = (float) (0.9 + (randdouble() * 0.1));
    float _b = 0;
 
    Particle::Add(new TractorAttractionParticle(pos, vec, _fade, _r, _g, _b, _target));
